+++ packages/react-native-reanimated/Common/cpp/reanimated/LayoutAnimations/SharedTransitions.cpp
@@ -1,8 +1,9 @@
 #include <folly/dynamic.h>
 #include <react/renderer/components/rnreanimated/Props.h>
 #include <react/renderer/components/scrollview/ScrollViewState.h>
-#include <reanimated/LayoutAnimations/LayoutAnimationsProxy_Experimental.h>
+#include <reanimated/LayoutAnimations/LayoutAnimationsProxy.h>
 #include <reanimated/LayoutAnimations/LayoutAnimationsUtils.h>
+#include <reanimated/Tools/FeatureFlags.h>
 #include <reanimated/Tools/ReanimatedSystraceSection.h>
 #include <ranges>
 
@@ -12,8 +13,7 @@ namespace reanimated {
 
 // A boundary is active when its `isActive` prop (controlled from JS,
 // e.g. with `useIsFocused`) is true and it's not currently exiting.
-std::shared_ptr<LightNode> LayoutAnimationsProxy_Experimental::findActiveBoundary(
-    const std::shared_ptr<LightNode> &node) const {
+std::shared_ptr<LightNode> LayoutAnimationsProxy::findActiveBoundary(const std::shared_ptr<LightNode> &node) const {
   std::shared_ptr<LightNode> result = nullptr;
 
   if (node->isExiting()) {
@@ -32,8 +32,7 @@ std::shared_ptr<LightNode> LayoutAnimationsProxy_Experimental::findActiveBoundar
   return result;
 }
 
-std::shared_ptr<LightNode> LayoutAnimationsProxy_Experimental::findBoundaryGuess(
-    const std::shared_ptr<LightNode> &node) const {
+std::shared_ptr<LightNode> LayoutAnimationsProxy::findBoundaryGuess(const std::shared_ptr<LightNode> &node) const {
   std::shared_ptr<LightNode> result = nullptr;
 
   if (node->isExiting()) {
@@ -52,7 +51,7 @@ std::shared_ptr<LightNode> LayoutAnimationsProxy_Experimental::findBoundaryGuess
   return result;
 }
 
-void LayoutAnimationsProxy_Experimental::findSharedElementsOnScreen(
+void LayoutAnimationsProxy::findSharedElementsOnScreen(
     const std::shared_ptr<LightNode> &node,
     BeforeOrAfter index,
     const PropsParserContext &propsParserContext,
@@ -97,7 +96,7 @@ void LayoutAnimationsProxy_Experimental::findSharedElementsOnScreen(
   }
 }
 
-void LayoutAnimationsProxy_Experimental::handleProgressTransition(
+void LayoutAnimationsProxy::handleProgressTransition(
     TransactionMeta &transaction,
     const ShadowViewMutationList &mutations,
     const PropsParserContext &propsParserContext) const {
@@ -196,7 +195,7 @@ void LayoutAnimationsProxy_Experimental::handleProgressTransition(
   }
 }
 
-void LayoutAnimationsProxy_Experimental::overrideTransform(
+void LayoutAnimationsProxy::overrideTransform(
     ShadowView &shadowView,
     const std::optional<Transform> &transform,
     const PropsParserContext &propsParserContext) const {
@@ -219,7 +218,7 @@ void LayoutAnimationsProxy_Experimental::overrideTransform(
   shadowView.props = newProps;
 }
 
-Tag LayoutAnimationsProxy_Experimental::getOrCreateContainer(
+Tag LayoutAnimationsProxy::getOrCreateContainer(
     const ShadowView &before,
     const SharedTag &sharedTag,
     TransactionMeta &transaction) const {
@@ -250,13 +249,13 @@ Tag LayoutAnimationsProxy_Experimental::getOrCreateContainer(
   return containerTag;
 }
 
-void LayoutAnimationsProxy_Experimental::handleSharedTransitionsStart(
+void LayoutAnimationsProxy::handleSharedTransitionsStart(
     const std::shared_ptr<LightNode> &afterTopScreen,
     const std::shared_ptr<LightNode> &beforeTopScreen,
     TransactionMeta &transaction,
     const ShadowViewMutationList &mutations,
     const PropsParserContext &propsParserContext) const {
-  ReanimatedSystraceSection s1("LayoutAnimationsProxy_Experimental::handleSharedTransitionsStart");
+  ReanimatedSystraceSection s1("LayoutAnimationsProxy::handleSharedTransitionsStart");
 
   if (!beforeTopScreen || !afterTopScreen) {
     return;
@@ -323,7 +322,7 @@ void LayoutAnimationsProxy_Experimental::handleSharedTransitionsStart(
   }
 }
 
-void LayoutAnimationsProxy_Experimental::hideTransitioningViews(
+void LayoutAnimationsProxy::hideTransitioningViews(
     BeforeOrAfter index,
     const Transitions &transitions,
     ShadowViewMutationList &mutations,
@@ -338,11 +337,11 @@ void LayoutAnimationsProxy_Experimental::hideTransitioningViews(
   }
 }
 
-std::optional<SurfaceId> LayoutAnimationsProxy_Experimental::onTransitionProgress(
-    int tag,
-    double progress,
-    bool isClosing,
-    bool isGoingForward) {
+std::optional<SurfaceId>
+LayoutAnimationsProxy::onTransitionProgress(int tag, double progress, bool isClosing, bool isGoingForward) {
+  if constexpr (!StaticFeatureFlags::getFlag("ENABLE_SHARED_ELEMENT_TRANSITIONS")) {
+    return {};
+  }
   auto lock = std::unique_lock<std::recursive_mutex>(mutex);
   const auto nodeIt = lightNodes_.find(tag);
   if (nodeIt == lightNodes_.end() || !nodeIt->second) {
@@ -373,7 +372,10 @@ std::optional<SurfaceId> LayoutAnimationsProxy_Experimental::onTransitionProgres
   return {};
 }
 
-std::optional<SurfaceId> LayoutAnimationsProxy_Experimental::onGestureCancel(int tag) {
+std::optional<SurfaceId> LayoutAnimationsProxy::onGestureCancel(int tag) {
+  if constexpr (!StaticFeatureFlags::getFlag("ENABLE_SHARED_ELEMENT_TRANSITIONS")) {
+    return {};
+  }
   auto lock = std::unique_lock<std::recursive_mutex>(mutex);
   const auto nodeIt = lightNodes_.find(tag);
   if (nodeIt == lightNodes_.end() || !nodeIt->second) {
@@ -387,7 +389,7 @@ std::optional<SurfaceId> LayoutAnimationsProxy_Experimental::onGestureCancel(int
   return {};
 }
 
-void LayoutAnimationsProxy_Experimental::insertContainers(TransactionMeta &transaction, int &rootChildCount) const {
+void LayoutAnimationsProxy::insertContainers(TransactionMeta &transaction, int &rootChildCount) const {
   auto &filteredMutations = transaction.filteredMutations;
   ShadowViewMutationList currentMutations;
   std::swap(currentMutations, filteredMutations);
@@ -400,12 +402,12 @@ void LayoutAnimationsProxy_Experimental::insertContainers(TransactionMeta &trans
   filteredMutations.insert(filteredMutations.end(), currentMutations.begin(), currentMutations.end());
 }
 
-void LayoutAnimationsProxy_Experimental::removeSharedContainer(Tag containerTag, TransactionMeta &transaction) const {
+void LayoutAnimationsProxy::removeSharedContainer(Tag containerTag, TransactionMeta &transaction) const {
   transaction.sharedContainersToRemove.push_back(containerTag);
   std::erase_if(containerTags_, [containerTag](const auto &entry) { return entry.second == containerTag; });
 }
 
-void LayoutAnimationsProxy_Experimental::cleanupSharedTransitions(
+void LayoutAnimationsProxy::cleanupSharedTransitions(
     TransactionMeta &transaction,
     const PropsParserContext &propsParserContext) const {
   ReanimatedSystraceSection s1("cleanupSharedTransitions");
@@ -443,7 +445,7 @@ void LayoutAnimationsProxy_Experimental::cleanupSharedTransitions(
 
 // MARK: Position Calculation
 
-std::vector<react::Point> LayoutAnimationsProxy_Experimental::getAbsolutePositionsForRootPathView(
+std::vector<react::Point> LayoutAnimationsProxy::getAbsolutePositionsForRootPathView(
     const std::shared_ptr<LightNode> &node) const {
   std::vector<react::Point> viewsAbsolutePositions;
   auto currentNode = node;
@@ -474,7 +476,7 @@ std::vector<react::Point> LayoutAnimationsProxy_Experimental::getAbsolutePositio
   return viewsAbsolutePositions;
 }
 
-std::optional<Transform> LayoutAnimationsProxy_Experimental::parseParentTransforms(
+std::optional<Transform> LayoutAnimationsProxy::parseParentTransforms(
     const std::shared_ptr<LightNode> &node,
     const std::vector<react::Point> &absolutePositions) const {
   std::vector<std::pair<Transform, TransformOrigin>> transforms;
@@ -529,7 +531,7 @@ std::optional<Transform> LayoutAnimationsProxy_Experimental::parseParentTransfor
 // https://github.com/facebook/react-native/blob/v0.80.0/packages/react-native/ReactCommon/react/renderer/components/view/BaseViewProps.cpp#L548
 // We need a copy of these methods to modify the `resolveTransform` method
 // to accept the transform origin as a parameter instead of as a class field.
-react::Transform LayoutAnimationsProxy_Experimental::resolveTransform(
+react::Transform LayoutAnimationsProxy::resolveTransform(
     const LayoutMetrics &layoutMetrics,
     const Transform &transform,
     const TransformOrigin &transformOrigin) const {
@@ -559,7 +561,7 @@ react::Transform LayoutAnimationsProxy_Experimental::resolveTransform(
   return transformMatrix;
 }
 
-std::array<float, 3> LayoutAnimationsProxy_Experimental::getTranslateForTransformOrigin(
+std::array<float, 3> LayoutAnimationsProxy::getTranslateForTransformOrigin(
     float viewWidth,
     float viewHeight,
     const TransformOrigin &transformOrigin) const {