+++ packages/react-native-reanimated/Common/cpp/reanimated/NativeModules/ReanimatedModuleProxy.cpp
@@ -13,7 +13,7 @@
 #include <reanimated/Events/UIEventHandler.h>
 #include <reanimated/Fabric/updates/PropsLayoutFilter.h>
 #include <reanimated/Fabric/updates/SynchronousPropNames.h>
-#include <reanimated/LayoutAnimations/LayoutAnimationsProxy_Experimental.h>
+#include <reanimated/LayoutAnimations/LayoutAnimationsProxy.h>
 #include <reanimated/LayoutAnimations/LayoutAnimationsProxy_Legacy.h>
 #include <reanimated/NativeModules/PropValueProcessor.h>
 #include <reanimated/NativeModules/ReanimatedModuleProxy.h>
@@ -1036,11 +1036,9 @@ bool ReanimatedModuleProxy::handleEventAndFlush(
 }
 
 void ReanimatedModuleProxy::applySynchronousUpdates(const UpdatesBatch &synchronousUpdatesBatch) {
-  if constexpr (StaticFeatureFlags::getFlag("ENABLE_SHARED_ELEMENT_TRANSITIONS")) {
-    if (layoutAnimationsProxyRegistry_ && !synchronousUpdatesBatch.empty() &&
-        DynamicFeatureFlags::getFlag("SYNCHRONOUS_PROPS_IN_LIGHT_TREE")) {
-      layoutAnimationsProxyRegistry_->applySynchronousProps(synchronousUpdatesBatch);
-    }
+  if (layoutAnimationsProxyRegistry_ && !synchronousUpdatesBatch.empty() &&
+      DynamicFeatureFlags::getFlag("SYNCHRONOUS_PROPS_IN_LIGHT_TREE")) {
+    layoutAnimationsProxyRegistry_->applySynchronousProps(synchronousUpdatesBatch);
   }
 
 #ifdef ANDROID
@@ -1263,8 +1261,8 @@ void ReanimatedModuleProxy::initializeLayoutAnimationsProxyRegistry() {
 #endif
   };
 
-  if constexpr (StaticFeatureFlags::getFlag("ENABLE_SHARED_ELEMENT_TRANSITIONS")) {
-    layoutAnimationsProxyRegistry_ = createLayoutAnimationsProxyExperimentalRegistry(dependencies);
+  if constexpr (!StaticFeatureFlags::getFlag("USE_LEGACY_LAYOUT_ANIMATIONS_PROXY")) {
+    layoutAnimationsProxyRegistry_ = createLayoutAnimationsProxyDefaultRegistry(dependencies);
   } else {
     layoutAnimationsProxyRegistry_ = createLayoutAnimationsProxyLegacyRegistry(dependencies);
   }